Incorpate feedback from PR #444
