+++ projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/LLVMPaths.java
@@ -76,8 +76,8 @@ public final class LLVMPaths {
     public static final String INTEROP_TESTS = LOCAL_TESTS + "/../interoptests";
 
     // LLVM test suite
-    static final File BCPARSER_TEST_SUITE_CONFIG = new File(EXTERNAL_TEST_SUITES_EXECUTION_CONFIG, "llvmbc/");
-    static final File BCPARSER_TEST_SUITE = new File(EXTERNAL_TEST_SUITES, "llvm/");
+    static final File LLVMBC_TEST_SUITE_CONFIG = new File(EXTERNAL_TEST_SUITES_EXECUTION_CONFIG, "llvmbc/");
+    static final File LLVMBC_TEST_SUITE = new File(EXTERNAL_TEST_SUITES, "llvm/");
 
     static {
         TEMP_DIRECTORY.mkdir();
